mint address input and wave object
